refactor : migrated code to mvvm architecture
