+++ app/build.gradle
@@ -52,6 +52,9 @@ dependencies {
     implementation 'androidx.recyclerview:recyclerview:1.1.0-alpha04'
     implementation 'androidx.cardview:cardview:1.0.0'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    //First Party Lifecycle Library
+    implementation "androidx.lifecycle:lifecycle-extensions:2.1.0-alpha04"
+    kapt "androidx.lifecycle:lifecycle-compiler:2.1.0-alpha04"
 
     //Material Design
     implementation 'com.google.android.material:material:1.1.0-alpha05'
