--- app/src/main/java/com/nilhcem/blenamebadge/data/ConfigInfo.kt
@@ -1,3 +1,0 @@
-package com.nilhcem.blenamebadge.data
-
-data class ConfigInfo(val badgeJSON: String, val fileName: String)
