+++ app/src/main/java/com/nilhcem/blenamebadge/data/badge_preview/Cell.kt
@@ -1,0 +1,5 @@
+package com.nilhcem.blenamebadge.data.badge_preview
+
+import android.graphics.Rect
+
+data class Cell(var list: MutableList<Rect> = mutableListOf())
