+++ app/src/main/java/com/nilhcem/blenamebadge/data/badge_preview/CheckList.kt
@@ -1,0 +1,7 @@
+package com.nilhcem.blenamebadge.data.badge_preview
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class CheckList(var list: MutableList<Boolean> = mutableListOf()) : Parcelable
