+++ app/src/main/java/com/nilhcem/blenamebadge/data/device/bluetooth/Constants.kt
@@ -1,0 +1,8 @@
+package com.nilhcem.blenamebadge.data.device.bluetooth
+
+import java.util.UUID
+
+object Constants {
+    val SERVICE_UUID: UUID = UUID.fromString("0000fee0-0000-1000-8000-00805f9b34fb")
+    val CHARACTERISTIC_UUID: UUID = UUID.fromString("0000fee1-0000-1000-8000-00805f9b34fb")
+}
