+++ app/src/main/java/com/nilhcem/blenamebadge/data/device/model/DataToSend.kt
@@ -1,0 +1,3 @@
+package com.nilhcem.blenamebadge.data.device.model
+
+data class DataToSend(val messages: List<Message>)
