+++ app/src/main/java/com/nilhcem/blenamebadge/data/device/model/Message.kt
@@ -1,0 +1,3 @@
+package com.nilhcem.blenamebadge.data.device.model
+
+data class Message(val hexStrings: List<String>, val flash: Boolean = false, val marquee: Boolean = false, val speed: Speed = Speed.ONE, val mode: Mode = Mode.LEFT)
