+++ app/src/main/java/com/nilhcem/blenamebadge/data/device/model/Speed.kt
@@ -1,0 +1,12 @@
+package com.nilhcem.blenamebadge.data.device.model
+
+enum class Speed(val hexValue: Byte) {
+    ONE(0x00),
+    TWO(0x10),
+    THREE(0x20),
+    FOUR(0x30),
+    FIVE(0x40),
+    SIX(0x50),
+    SEVEN(0x60),
+    EIGHT(0x70)
+}
