+++ app/src/main/java/com/nilhcem/blenamebadge/data/fragments/ConfigInfo.kt
@@ -1,0 +1,3 @@
+package com.nilhcem.blenamebadge.data.fragments
+
+data class ConfigInfo(val badgeJSON: String, val fileName: String)
