+++ app/src/main/java/com/nilhcem/blenamebadge/data/fragments/FilesContainer.kt
@@ -1,0 +1,18 @@
+package com.nilhcem.blenamebadge.data.fragments
+
+class FilesContainer private constructor() {
+
+    var filesData = FilesDAO()
+        private set
+
+    companion object {
+        @Volatile
+        private var instance: FilesContainer? = null
+
+        fun getInstance() =
+            instance ?: synchronized(this) {
+                instance
+                    ?: FilesContainer().also { instance = it }
+            }
+    }
+}
