+++ app/src/main/java/com/nilhcem/blenamebadge/data/util/SendingData.kt
@@ -1,0 +1,6 @@
+package com.nilhcem.blenamebadge.data.util
+
+import com.nilhcem.blenamebadge.data.device.model.Mode
+import com.nilhcem.blenamebadge.data.device.model.Speed
+
+data class SendingData(val invertLED: Boolean, val flash: Boolean, val marquee: Boolean, val mode: Mode, val speed: Speed)
