--- app/src/main/java/com/nilhcem/blenamebadge/device/model/DataToSend.kt
@@ -1,3 +1,0 @@
-package com.nilhcem.blenamebadge.device.model
-
-data class DataToSend(val messages: List<Message>)
