--- app/src/main/java/com/nilhcem/blenamebadge/ui/badge_preview/Cell.kt
@@ -1,5 +1,0 @@
-package com.nilhcem.blenamebadge.ui.badge_preview
-
-import android.graphics.Rect
-
-data class Cell(var list: MutableList<Rect> = mutableListOf())
