--- app/src/main/java/com/nilhcem/blenamebadge/ui/badge_preview/CheckList.kt
@@ -1,7 +1,0 @@
-package com.nilhcem.blenamebadge.ui.badge_preview
-
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-
-@Parcelize
-data class CheckList(var list: MutableList<Boolean> = mutableListOf()) : Parcelable
