--- app/src/main/java/com/nilhcem/blenamebadge/ui/main/MainPresenter.kt
@@ -1,52 +1,0 @@
-package com.nilhcem.blenamebadge.ui.main
-
-import android.content.Context
-import android.widget.Toast
-import com.nilhcem.blenamebadge.R
-import com.nilhcem.blenamebadge.core.android.log.Timber
-import com.nilhcem.blenamebadge.core.utils.ByteArrayUtils
-import com.nilhcem.blenamebadge.device.DataToByteArrayConverter
-import com.nilhcem.blenamebadge.device.bluetooth.GattClient
-import com.nilhcem.blenamebadge.device.bluetooth.ScanHelper
-import com.nilhcem.blenamebadge.device.model.DataToSend
-
-class MainPresenter {
-
-    private val scanHelper = ScanHelper()
-    private val gattClient = GattClient()
-
-    fun sendMessage(context: Context, dataToSend: DataToSend) {
-        Timber.i { "About to send data: $dataToSend" }
-        val byteData = DataToByteArrayConverter.convert(dataToSend)
-        sendBytes(context, byteData)
-    }
-
-    fun onPause() {
-        scanHelper.stopLeScan()
-        gattClient.stopClient()
-    }
-
-    private fun sendBytes(context: Context, byteData: List<ByteArray>) {
-        Timber.i { "ByteData: ${byteData.map { ByteArrayUtils.byteArrayToHexString(it) }}" }
-
-        scanHelper.startLeScan { device ->
-            if (device == null) {
-                Timber.e { "Scan could not find any device" }
-                Toast.makeText(context, R.string.no_device_found, Toast.LENGTH_SHORT).show()
-            } else {
-                Timber.e { "Device found: $device" }
-
-                gattClient.startClient(context, device.address) { onConnected ->
-                    if (onConnected) {
-                        gattClient.writeDataStart(byteData) {
-                            Timber.i { "Data sent" }
-                            gattClient.stopClient()
-                        }
-                    } else {
-                        Toast.makeText(context, R.string.no_device_found, Toast.LENGTH_SHORT).show()
-                    }
-                }
-            }
-        }
-    }
-}
