+++ app/src/main/java/com/nilhcem/blenamebadge/util/StorageUtils.kt
@@ -3,13 +3,13 @@ package com.nilhcem.blenamebadge.util
 import android.content.Context
 import android.net.Uri
 import android.os.Environment
-import com.nilhcem.blenamebadge.data.ConfigInfo
-import com.nilhcem.blenamebadge.data.CONF_FLASH
-import com.nilhcem.blenamebadge.data.CONF_HEX_STRINGS
-import com.nilhcem.blenamebadge.data.CONF_MARQUEE
-import com.nilhcem.blenamebadge.data.CONF_INVERTED
-import com.nilhcem.blenamebadge.data.CONF_MODE
-import com.nilhcem.blenamebadge.data.CONF_SPEED
+import com.nilhcem.blenamebadge.data.fragments.CONF_HEX_STRINGS
+import com.nilhcem.blenamebadge.data.fragments.CONF_INVERTED
+import com.nilhcem.blenamebadge.data.fragments.CONF_MARQUEE
+import com.nilhcem.blenamebadge.data.fragments.CONF_FLASH
+import com.nilhcem.blenamebadge.data.fragments.CONF_MODE
+import com.nilhcem.blenamebadge.data.fragments.CONF_SPEED
+import com.nilhcem.blenamebadge.data.fragments.ConfigInfo
 import org.json.JSONObject
 import java.io.File
 import java.io.BufferedReader
